GH-11: Replace 'pnpm' with 'npm' and remove TinaCMS integrations.
this close #11
Migrated the CMS project from 'pnpm' to 'npm', simplifying dependency handling and development processes. Removed unused TinaCMS configurations, workflows, and scripts, streamlining the project setup. Updated package-lock and introduced standard Busybox configs for linting, formatting, and TypeScript.
