+++ lerna.json
@@ -1,0 +1,6 @@
+{
+  "$schema": "node_modules/lerna/schemas/lerna-schema.json",
+  "npmClient": "npm",
+  "packages": ["systems/*"],
+  "version": "2024.12.2-alpha.1524"
+}
