+++ package.json
@@ -9,6 +9,7 @@
     "eslint": "9.15.0",
     "globals": "15.12.0",
     "husky": "9.1.7",
+    "lerna": "8.1.9",
     "lint-staged": "15.2.10",
     "prettier": "3.3.3"
   },
