--- scripts/ci/setup.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-set -ex
-
-npm install
