+++ systems/cms/.prettierrc.mjs
@@ -1,0 +1,3 @@
+import busyboxPrettierConfig from "@busybox/prettier-config";
+
+export default busyboxPrettierConfig;
