+++ systems/cms/README.md
@@ -1,3 +1,5 @@
+# @portfoilo/cms
+
 This is a [Tina CMS](https://tina.io/) project.
 
 ## Local Development
@@ -5,16 +7,20 @@ This is a [Tina CMS](https://tina.io/) project.
 Install the project's dependencies:
 
 > [!NOTE]  
-> [Do you know the best package manager for Node.js?](https://www.ssw.com.au/rules/best-package-manager-for-node/) Using the right package manager can greatly enhance your development workflow. We recommend using pnpm for its speed and efficient handling of dependencies. Learn more about why pnpm might be the best choice for your projects by checking out this rule from SSW.
+> [Do you know the best package manager for Node.js?](https://www.ssw.com.au/rules/best-package-manager-for-node/) Using
+> the right package manager can greatly enhance your development workflow. We recommend using pnpm for its speed and
+> efficient handling of dependencies. Learn more about why pnpm might be the best choice for your projects by checking
+> out
+> this rule from SSW.
 
-```
-pnpm install
+```bash
+npm install
 ```
 
 Run the project locally:
 
-```
-pnpm dev
+```bash
+npm run dev
 ```
 
 Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
@@ -23,7 +29,7 @@ Open [http://localhost:3000](http://localhost:3000) with your browser to see the
 
 Replace the `.env.example`, with `.env`
 
-```
+```bash
 NEXT_PUBLIC_TINA_CLIENT_ID=<get this from the project you create at app.tina.io>
 TINA_TOKEN=<get this from the project you create at app.tina.io>
 NEXT_PUBLIC_TINA_BRANCH=<Specify the branch with Tina configured>
@@ -32,7 +38,7 @@ NEXT_PUBLIC_TINA_BRANCH=<Specify the branch with Tina configured>
 Build the project:
 
 ```bash
-pnpm build
+npm run build
 ```
 
 ## Learn More
@@ -42,6 +48,7 @@ To learn more about Tina, take a look at the following resources:
 - [Tina Docs](https://tina.io/docs)
 - [Getting started](https://tina.io/docs/setup-overview/)
 
-You can check out [Tina Github repository](https://github.com/tinacms/tinacms) - your feedback and contributions are welcome!
+You can check out [Tina Github repository](https://github.com/tinacms/tinacms) - your feedback and contributions are
+welcome!
 
 ## [Deploy on Vercel](https://tina.io/guides/tina-cloud/add-tinacms-to-existing-site/deployment/)
