+++ systems/cms/eslint.config.mjs
@@ -1,14 +1,14 @@
-import { useCodeSortingEslintConfig } from '@busybox/eslint-config-code-sorting';
-import { useESModuleEslintConfig } from '@busybox/eslint-config-esm';
+import { useCodeSortingEslintConfig } from "@busybox/eslint-config-code-sorting";
+import { useESModuleEslintConfig } from "@busybox/eslint-config-esm";
 import {
   useJSONEslintConfig,
   useMarkdownEslintConfig,
   usePackageJsonEslintConfig,
   useYamlEslintConfig,
-} from '@busybox/eslint-config-text-document';
-import globals from 'globals';
+} from "@busybox/eslint-config-text-document";
+import globals from "globals";
 
-import pkgjson from './package.json' with { type: 'json' };
+import pkgjson from "./package.json" with { type: "json" };
 
 function withOverride(overrideConfig) {
   return function createEslintConfigHOC(originalConfigFunc) {
@@ -28,7 +28,11 @@ function withOverride(overrideConfig) {
 
 export default [
   {
-    ignores: ['package-lock.json', 'systems/**/*'],
+    ignores: [
+      "package-lock.json",
+      "tina/tina-lock.json",
+      "tina/__generated__/**/*",
+    ],
     name: pkgjson.name,
   },
   {
@@ -44,7 +48,7 @@ export default [
   useJSONEslintConfig(),
   withOverride({
     rules: {
-      'markdownlint/md013': 'off',
+      "markdownlint/md013": "off",
     },
   })(useMarkdownEslintConfig)(),
 ];
