+++ systems/cms/scripts/dev-setup.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+npm install
