+++ systems/cms/scripts/dev.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+tinacms dev
