+++ systems/cms/tina/collections/post.ts
@@ -1,26 +1,26 @@
-/**
- * @type {import('tinacms').Collection}
- */
-export default {
+import type { Collection } from "tinacms";
+
+const Posts: Collection = {
+  fields: [
+    {
+      label: "Title",
+      name: "title",
+      type: "string",
+    },
+    {
+      isBody: true,
+      label: "Blog Post Body",
+      name: "body",
+      type: "rich-text",
+    },
+  ],
   label: "Blog Posts",
   name: "post",
   path: "content/post",
-  fields: [
-    {
-      type: "string",
-      label: "Title",
-      name: "title",
-    },
-    {
-      type: "rich-text",
-      label: "Blog Post Body",
-      name: "body",
-      isBody: true,
-    },
-  ],
   ui: {
     router: ({ document }) => {
       return `/posts/${document._sys.filename}`;
     },
   },
 };
+export default Posts;
